+++ SampleTCPClient/src/main/java/de/bundeswehr/mese/sedapexpress/sample/tcpclient/SampleTCPClient.java
@@ -48,7 +48,7 @@ public class SampleTCPClient implements SEDAPExpressSubscriber {
 
     private final String senderId;
 
-    private short numberSTATUS = 0;
+    private byte numberSTATUS = 0;
 
     /**
      * Instantiate a sample TCP client
